src/main/java/com/gudimov/movieland/dao/jdbc/mapper/UserRowMapper.java
package com.gudimov.movieland.dao.jdbc.mapper;

import com.gudimov.movieland.entity.User;
import org.springframework.jdbc.core.RowMapper;

import java.sql.ResultSet;
import java.sql.SQLException;

public class UserRowMapper implements RowMapper<User> {
    @Override
    public User mapRow(ResultSet resultSet, int i) throws SQLException {
        User user = new User();
        user.setId(resultSet.getInt("id"));
        user.setNickName(resultSet.getString("nickname"));
        return user;
    }
}
